优化删除专业信息相关业务逻辑

+++ course/src/main/java/com/jiangzhichao/service/admin/impl/AdminOpDepartmentServiceImpl.java
@@ -6,8 +6,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.jiangzhichao.dao.CourseDOMapper;
 import com.jiangzhichao.dao.DepartmentDOMapper;
+import com.jiangzhichao.dao.StudentDOMapper;
+import com.jiangzhichao.entity.CourseDO;
 import com.jiangzhichao.entity.DepartmentDO;
+import com.jiangzhichao.entity.StudentDO;
 import com.jiangzhichao.service.admin.AdminOpDepartmentService;
 
 @Service
@@ -17,6 +21,12 @@ public class AdminOpDepartmentServiceImpl implements AdminOpDepartmentService {
 	@Autowired
 	private DepartmentDOMapper departmentDOMapper;
 	
+	@Autowired
+	private CourseDOMapper courseDOMapper;
+	
+	@Autowired
+	private StudentDOMapper studentDOMapper;
+	
 	@Override
 	public int insertDepartment(DepartmentDO departmentDO) {
 		return departmentDOMapper.insert(departmentDO);
@@ -24,8 +34,17 @@ public class AdminOpDepartmentServiceImpl implements AdminOpDepartmentService {
 
 	@Override
 	public int deleteDepartment(String dno) {
-		//删除前先判断此专业下是否有学生
-		int i = departmentDOMapper.deleteByPrimaryKey(dno);
+		int i = 0;
+		//判断此专业下是否有课程
+		List<CourseDO> courses = courseDOMapper.selectByDno(dno);
+		if(courses.size() == 0) {
+			//删除前先判断此专业下是否有学生
+			List<StudentDO> students = studentDOMapper.selectByDno(dno);
+			if(students.size() == 0) {
+				//删除专业信息
+				i = departmentDOMapper.deleteByPrimaryKey(dno);
+			}
+		}
 		return i;
 	}
